Lab3/pzpi-22-3-bilozub-danylo-lab3/ZooShopHelper/requirements.txt
anyio==3.6.2
async-timeout==4.0.3
asyncpg==0.29.0
bcrypt==4.2.0
certifi==2024.8.30
click==8.1.3
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.95.2
fastapi-jwt-auth==0.5.0
fastapi-utils==0.8.0
fonttools==4.51.0
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
mpmath==1.3.0
numpy==1.26.4
orjson==3.10.7
packaging==24.0
passlib==1.7.4
pillow==10.3.0
psutil==5.9.8

psycopg2-binary
pyasn1==0.6.1
pydantic==1.10.8
PyJWT==1.7.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.2
rsa==4.9
schedule==1.2.2
scikit-learn==1.4.2
scipy==1.13.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.34
starlette==0.27.0
sympy==1.12
threadpoolctl==3.5.0
typing_extensions==4.6.2
ujson==5.10.0
uvicorn==0.22.0
watchfiles==0.24.0
websockets==13.0.1
